temporary fix for probing

+++ nifty6/probing.py
@@ -97,10 +97,12 @@ def probe_with_posterior_samples(op, post_op, nprobes):
             raise ValueError
     sc = StatCalculator()
     for i in range(nprobes):
+        # FIXME which dtype should we use here?
+        import numpy as np
         if post_op is None:
-            sc.add(op.draw_sample(from_inverse=True))
+            sc.add(op.draw_sample(dtype=np.float64, from_inverse=True))
         else:
-            sc.add(post_op(op.draw_sample(from_inverse=True)))
+            sc.add(post_op(op.draw_sample(dtype=np.float64, from_inverse=True)))
 
     if nprobes == 1:
         return sc.mean, None
